Movies Table & SearchBar finish

+++ dxc/dxc/core/models/Movie.swift
@@ -11,7 +11,7 @@ struct Movie: Codable {
     let id: Int
     let original_title: String
     let overview: String
-    let vote_average: Decimal
+    let vote_average: Float
     let poster_path: String
     
 }
